backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email]==2.7.4
pydantic-settings==2.3.0
pillow==10.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
pymongo==4.8.0
kafka-python==2.0.2
langchain==0.2.5
langchain-core==0.2.43
langchain-openai==0.1.8
langchain-community==0.2.5
tavily-python==0.3.3
openai==1.35.3
httpx==0.27.2
langchain-tavily
